Generator

Description

Generates usable data for insert statements

Documentation

data PSQLTYPE

Describes a few common types used in PSQL

PSQLTYPE constructors

showPSQLTYPE :: PSQLTYPE -> String

Used in the PSQLTYPE instance of Show to print

psqlVarchar :: String -> PSQLTYPE

Convert string to PSQLTYPE of VARCHAR String

unVarchar :: PSQLTYPE -> String

Convert VARCHAR to String

psqlDate :: (Int, Int, Int) -> PSQLTYPE

Convert a tuple of (year,month,day) to PSQLTYPE Date

unDate :: PSQLTYPE -> (Int, Int, Int)

Convert a PSQLTYPE Date to a tuple (year,month,day)

psqlInteger :: Int -> PSQLTYPE

Convert Int to PSQLTYPE INTEGER

unInteger :: PSQLTYPE -> Int

Convert PSQLTYPE INTEGER to Int

psqlTimestamp :: (Int, Int, Int) -> (Int, Int, Int) -> Int -> Bool -> PSQLTYPE

Convert tuples (year,month,day) (hour,minute,second) to PSQLTYPE TIMESTAMP

Attribute gen combinators

make :: Int -> Gen a -> Gen [a]

Make an n long list of randomly generated items

unique :: Eq a => Int -> Gen a -> Gen [a]

Create a list of unique generated items of length n

Attribute generators

psqlName :: PSQLTYPE -> PSQLTYPE -> PSQLTYPE

Generate a full name based on a first and last name in PSQLTYPE Varchar

firstnames :: Int -> Gen [PSQLTYPE]

Generate a list of firstnames

lastnames :: Int -> Gen [PSQLTYPE]

Generate a list of lastnames

primaryKeys :: Int -> Gen [PSQLTYPE]

Produces a list of unique primary keys at the specified length

name :: Gen PSQLTYPE

generates a random full name

name2 :: PSQLTYPE -> PSQLTYPE -> PSQLTYPE

Generates a full name using the given first and lastname

email :: Gen PSQLTYPE

Generates a random email

email2 :: PSQLTYPE -> PSQLTYPE -> Gen PSQLTYPE

Generates an email with the specified first and lastname

date :: Int -> Gen PSQLTYPE

Creates a date with the specified year

dateBetween :: (Int, Int, Int) -> (Int, Int, Int) -> Gen PSQLTYPE

Generate a date between two dates of the form (year,month,day)

timestampBetween :: (Int, Int, Int) -> (Int, Int, Int) -> Gen PSQLTYPE

Provides a random timestamp between two dates, the time is random though

eventTimestampBetween :: (Int, Int, Int) -> (Int, Int, Int) -> Gen (PSQLTYPE, PSQLTYPE)

Generates two timestamps on the same day but hours between each other

eventTitle :: Gen PSQLTYPE

Generate a random event title

url :: String -> PSQLTYPE -> Gen PSQLTYPE

Generate a random URL of the form "http://kthsocial.com/S/i*i" where S is the string passed and i is the PSQLTYPE Integer passed

tagList :: Gen PSQLTYPE

Generate a list of hashtags

place :: Gen PSQLTYPE

Generate a random address

imageFilter :: Gen PSQLTYPE

Generate a random image filter

Text generators

imfeeling :: Gen PSQLTYPE

Generate random text of the form "I'm feeling emotion"

makesmefeel :: Gen PSQLTYPE

Generate random text of the form "That makes me feel emotion"

simpleSentance :: Gen PSQLTYPE

Generate random text of the form phrase predicate preposition nounphrase, high risk of sounding unnatural.

simpleSentance2 :: Gen PSQLTYPE

Generate random text of the form phrase phrase nounphrase, sentances are generally of acceptable grammar.

feelabout :: Gen PSQLTYPE

Generate random Text of the form "I'm feeling emotion about the event title event"

tweet :: Gen PSQLTYPE

Choose a random tweet from tweets in hs

sentanceAnd :: Gen PSQLTYPE -> Gen PSQLTYPE -> Gen PSQLTYPE

Combine sentance generators into paragraphs.

text :: Gen PSQLTYPE

Generate random text from several sentance generators. Risk of wonky grammar.

goodText :: Gen PSQLTYPE

Generate random text from several sentance generators. Generally acceptable grammar