Populator

Documentation

data InsertStatement

Represents an Insert Statement in postgreSequal

Constructors

IS

Fields

schemaName :: String

attributes :: [String]

values :: [PSQLTYPE]

Statements [InsertStatement]

Insert Statement constructors

insertStatement :: String -> [String] -> [PSQLTYPE] -> InsertStatement

Creates an InsertStatement object based on a schemaName, list of attributes and list of PSQLTYPE values for each attribute.

insert :: String -> [String] -> [[PSQLTYPE]] -> InsertStatement

Creates many InsertStatement objects in a collection, using the shared schemaName, shared attributes and a list of lists of values. E.g. insert "user" ["name","email"] [[names],[emails]]

Statement Generators

selfRefPairs :: Int -> [PSQLTYPE] -> Gen ([Int], [[PSQLTYPE]])

Create non reflexive pairs from one list of key. For each key, pair it with unique keys from the same list. Each key gets a maximum nr of random pairings.

selfRefPairs' :: Int -> [PSQLTYPE] -> Gen [[PSQLTYPE]]

Same ass selfRedPairs but without the nr of pairs per element.

pairs2' :: [PSQLTYPE] -> [PSQLTYPE] -> Gen [(PSQLTYPE, PSQLTYPE)]

For each element in the first list, choose a random element from the second list. Elements from the first list occur only once, but elements from the second list occur multiple times.

forEachKeyMakePairs :: (Int, Int) -> [PSQLTYPE] -> [PSQLTYPE] -> Gen ([Int], [[PSQLTYPE]])

For each key in the first list choose up to n random and unique keys in the second list.

forEachKeyMakePairs' :: (Int, Int) -> [PSQLTYPE] -> [PSQLTYPE] -> Gen [[PSQLTYPE]]

Same as forEachKeyMakePairs but discards the nr of pairs per key list.

pairFactory :: (Int, Int) -> [PSQLTYPE] -> [PSQLTYPE] -> ([PSQLTYPE] -> (PSQLTYPE, Int, Int) -> Gen (Int, [(PSQLTYPE, PSQLTYPE)])) -> Gen ([Int], [[PSQLTYPE]])

Make pairs based on a function that uses a list to choose from, the current element to pair to, how many to pair it to and what index it has. Returns the number of pairs per element from the first list and and two lists of an unziped pair in a list. ([nr of pairs per element],[[p1],[p2]]). Used in forEachKeyMakePairs, and selfRefPairs.

forEachDateMakeDates :: [PSQLTYPE] -> [Int] -> Gen [[PSQLTYPE]]

For each date in the list of dates, make a date that is equal to or later than that date.