Pretty

Documentation

pretty :: Gen InsertStatement -> IO ()

Pretty print a Gen InsertStatement in the PSQL format of insert statements. Can either take a collection of InsertStatement or a single one.

debugPSQLList :: Foldable t => t -> (t -> Gen (t PSQLTYPE)) -> IO ()

Takes an n and a list generator of random PSQLTYPE data and prints it in a readible format for debuging.

debugPSQL :: Gen PSQLTYPE -> IO ()

Takes a generator and prints it in a readible format for debuging

debugPSQL' :: Int -> Gen PSQLTYPE -> IO ()

Takes a generator and the amount of data to be generated, prints that amount in a readible format for debuging.